requirements.txt
#pip install -r requirements.txt

python-dotenv
requests
openai
